Perf: Try solveByElim later (esp. the And version).
This requires a refactored check that all facts are used since
anonymous lemmas can be used both to check given facts and for the final
proof. We use a heuristic: if a given fact is not used in the final
proof but the lemma used to prove it is used they we don’t complain.